+++ src/rest/defaults.ts
@@ -1,12 +1,14 @@
 import {Type} from '@angular/core';
 
 import {RestMiddleware} from './rest.interface';
+import {ReportProgressMiddleware} from './middlewares';
 
 /**
  * Definition of basic default array of rest middlewares order
  */
 export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
 [
+    ReportProgressMiddleware
 ];
 
 /**
@@ -14,4 +16,5 @@ export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
  */
 export const BASIC_DEFAULT_REST_METHOD_MIDDLEWARES: Type<RestMiddleware>[] =
 [
+    ReportProgressMiddleware
 ];