+++ src/rest/middlewares/index.ts
@@ -1,1 +1,2 @@
-export * from './parameters.middleware';
+// export * from './parameters.middleware';
+export * from './reportProgress.middleware';